+++ src/app/internal/income/other-income/income-list.component.ts
@@ -7,10 +7,10 @@ import moment from "moment";
 import {State} from "../../../core/state";
 import {Router} from "@angular/router";
 import {NumberPipe} from "../../../lib/ffc-angular/pipes/number.pipe";
-import {ModalService} from "../../../lib/ffc-angular/services/modal.service";
 import {TableOptions} from '../../../lib/ffc-angular/components/table/table-options.model';
 import {SortDirection} from '../../../lib/ffc-angular/components/table/sort-direction.enum';
 import {ColumnAlignment} from '../../../lib/ffc-angular/components/table/column-alignment.enum';
+import {DialogService, DialogType} from '@freshfox/ng-core';
 
 @Component({
 	templateUrl: './income-list.component.html'
@@ -34,7 +34,7 @@ export class IncomeListComponent implements OnInit {
 				private datePipe: DatePipe,
 				private state: State,
 				private router: Router,
-				private modalService: ModalService) {
+				private dialogService: DialogService) {
 	}
 
 	ngOnInit() {
@@ -102,19 +102,20 @@ export class IncomeListComponent implements OnInit {
 	}
 
 	deleteIncome(income: Income) {
-		this.modalService.createConfirmRequest(
+		this.dialogService.createConfirmRequest(
 			this.translate.instant('income.delete-confirm-title'),
 			this.translate.instant('income.delete-confirm-message'),
-			() => {
-				this.modalService.hideCurrentModal();
-			},
-			() => {
+			null,
+			null,
+			DialogType.Danger
+		).subscribe(confirmed => {
+			if (confirmed) {
 				let index = this.incomes.indexOf(income);
 				this.incomes.splice(index, 1);
 				this.filter();
 				this.incomeService.deleteIncome(income).subscribe();
-				this.modalService.hideCurrentModal();
-			});
+			}
+		});
 	}
 
 }
